feat: User Service 연동 리팩토링
- User Service와 연동 후 유저 정보를 가져오는 부분을 공통으로 사용하기에 해당 기능을 하나의 클래스가 담당하도록 수정

Related to: #CO-216
