--- meeting-service/src/main/java/com/comeon/meetingservice/web/common/feign/userservice/UserStatus.java
@@ -1,7 +1,0 @@
-package com.comeon.meetingservice.web.common.feign.userservice;
-
-public enum UserStatus {
-
-    ACTIVATE,
-    WITHDRAWN;
-}
