+++ meeting-service/src/main/java/com/comeon/meetingservice/web/common/feign/userservice/response/UserStatus.java
@@ -1,0 +1,7 @@
+package com.comeon.meetingservice.web.common.feign.userservice.response;
+
+public enum UserStatus {
+
+    ACTIVATE,
+    WITHDRAWN;
+}
